# Return undefined from graphql-js visit when not modifying node