--- a/src/transforms/applyDefaultNullability.ts
+++ b/src/transforms/applyDefaultNullability.ts
@@ -54,7 +54,7 @@ export function applyDefaultNullability(
         }
         return { ...t, directives, type };
       }
-      return t;
+      return undefined;
     },
   });
   if (errors.length > 0) {