Graphql-js has a faster path if you return undefined. Without that, it will recreate the parent object and re-assign the unmodified as if it were a new node.

To profile I used our new profile tool which creates a project with 10,000 types.

## Before

Grats completed in: 9.954s
Grats completed in: 9.967s
Grats completed in: 9.578s
Grats completed in: 9.609s

## After

Grats completed in: 9.090s
Grats completed in: 8.969s
Grats completed in: 9.007s
Grats completed in: 9.035s

This saves us on the order of 5-10%